macOS Accessibility: add an elementWithInterface helper
Instead of explicitly creating an ID from an interface and then asking
for the element for the ID, ask for the element for the interface
directly.

In that helper we can also make sure that the created element is
correctly configured if the interface for which it was created was for a
table cell.

Task-number: QTBUG-34337
Pick-to: 6.5
Change-Id: Id0f9247b0b50195301b293dcabb8925c3fc2d2cf
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>