Revert "macOS: add default Edit menu items, if not added by AppKit"
This reverts commit d42cfeb84faf154b46f2811b2059946b396fcc12.

Trying to help AppKit by adding the dictation and emoji edit menu items
ourselves resulted in sometimes adding the entries twice, when we failed
to detect the existing entries, or AppKit failed to detect our entries
as a reason to not add its own.

In addition, even if the entries we added were detected by AppKit and
AppKit was smart enough to not add its own, our entries were relying
on the developer to provide translation, instead of building on the
translations that AppKit already provides. And the keyboard shortcut
we set for our entries were not following system and user preferences
for which keyboard combination should trigger the entries.

Fixes: QTBUG-104709
Task-number: QTBUG-79565
Change-Id: I3fabc41f85df917dbb669253ad441bccea8a5e35
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit 9618427ee5a42e1d120fa475aa6e2408a738915a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>