FontDatabase[win/freetype] implement population of aliases
We rely on this function discovering extra fonts if we cannot find any
or only find poor matches for our font request. Whether that be because
we had only partially populated the fonts of a given family or because
the user specified a font using a localized name.

It's a copy-paste of the implementation in qwindowsfontdatabase

Fixes: QTBUG-111854
Change-Id: I0dd97b68d84417a26ce82a46f39804b36602ed67
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit 7a92304040c375fe867f71c1b04bc3e6b2c32b6e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>