qmake: Don't add dummy empty.lproj localization to macOS bundles
The original change from 2008 mentions that this was needed for
preventing a crash when searching the native file dialog, but
this has since been fixed.

Nor is the file needed to get localized native file dialogs,
as this is controlled by the combination of the app's supported
localization or CFBundleAllowMixedLocalizations.

And we don't do this for CMake projects.

Change-Id: I3c9e5aee4707c019f733920eb088f8d84f8e4ee1
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit f2b59f313857ef4834835976645b083e51cd15d2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>