QIcon: avoid upscaling icons if a higher resolution is available
The freedesktop spec [1] defines an algorithm to select a matching icon
by comparing the distance to the nearest resolution. So when having a
8x8 icon and a 8x8@2 icon the algorithm prefers the 8x8 icon if we
request an 8x8@1.25. Changing the algorithm to operate on a delta allows
us to check whether the icon is bigger or smaller than requested, and
with that information we can make sure to select the icon with the
closest size that is bigger than requested.

[1] https://specifications.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html

Task-number: QTBUG-90634
Change-Id: I116f161f9382b618c723ccf1653ce3bac0d79240
Reviewed-by: David Faure <david.faure@kdab.com>